WoT restructure B3 fixup: update test imports to axisSizing/ (positionCache, rowUtils, columnUtils)
