+++ handsontable/src/3rdparty/walkontable/src/settings/accessor.ts
