+++ handsontable/src/3rdparty/walkontable/test/unit/calculator/renderAllColumns.unit.ts
@@ -2,7 +2,7 @@ import {
   ViewportColumnsCalculator,
   RenderedAllColumnsCalculationType,
 } from '../../../src/calculator';
-import { PositionCache } from '../../../src/utils/positionCache';
+import { PositionCache } from '../../../src/axisSizing/positionCache';
 
 /**
  *
