+++ handsontable/src/3rdparty/walkontable/test/unit/calculator/renderAllRows.unit.ts
@@ -2,7 +2,7 @@ import {
   ViewportRowsCalculator,
   RenderedAllRowsCalculationType,
 } from '../../../src/calculator';
-import { PositionCache } from '../../../src/utils/positionCache';
+import { PositionCache } from '../../../src/axisSizing/positionCache';
 
 /**
  *
